Move pdf helper to DonationApplication#export_pdf
Move DonationApplicationsController#export_pdf to DonationApplication#export_pdf since it handles responsibilities that belong to the model, like knowing its own contents. Inject dependency (pdf object) to the controller methods DonationApplicationsController#display_pdf and DonationApplicationsController#download_pdf to retrieve the form from the model and pass it to the browser.